third commit

+++ Controllers/teacherControllers.js
@@ -5,6 +5,9 @@ const hodModel=require("../models/hodModel.js")
 const teacherModel=require("../models/teacherModel.js")
 const sectionModel=require("../models/sectionModel.js")
 const notesModel=require("../models/notesModel.js")
+const subjectModel=require("../models/subjectModel.js")
+const meetingModel=require("../models/meetingsModel.js")
+const studentModel=require("../models/studentModel.js")
 const date = require('date-and-time');
 
 const {SECRET_KEY}=require("../secrets.js")
@@ -22,6 +25,12 @@ exports.signup=async(req,res)=>{
         }
         
         const collegeId=data.collegeId;
+        
+        const emailExists=await teacherModel.findOne({email:data.email})
+        if(emailExists)
+        {
+            throw new Error("Email Exists")
+        }
         
         const college=await principalModel.findOne({collegeId:collegeId});
         
@@ -219,20 +228,34 @@ exports.showteacherNotes=async(req,res)=>
 
 
 exports.deleteNotes=async(req,res)=>{
-    
-    
+
     try{
         
         const notesId=req.params.id;
-        console.log(notesId)
-        await notesModel.deleteOne({_id:notesId});
+        // console.log(notesId)
         
-        res.status(201).json({
+        const notes=await notesModel.findById(notesId);
+        console.log(notes.teacher)
+        console.log(req.teacher._id)
+        const tid=String(req.teacher._id);
+        const ntid=String(notes.teacher);
+        
+        if(tid===ntid)
+        {
+        await notesModel.deleteOne({_id:notesId});    
+         res.status(201).json({
            status:"success",
            data:"notes deleted Successfully"
            
              })
+        }else{
         
+        
+        res.status(500).json({
+            status:"failure",
+            error:"Your Not Authorized to delete it "
+        }) 
+        }
         
     }
     catch(err)
@@ -252,8 +275,29 @@ exports.addSubject=async(req,res)=>{
     
     
     try{
-        console.log(req.body);
         
+        const subjectFound=await subjectModel.findOne({name:req.body.name,year:req.body.year});
+        
+        const dId=req.teacher.department;
+        
+        if(subjectFound)
+        {
+            throw new Error("Already subject added");
+        }
+        
+
+        const subject=new subjectModel(req.body);
+        
+        subject.department=dId;
+        
+        await subject.save();
+        
+        
+        res.status(201).json({
+           status:"success",
+           data:subject
+           
+             })
         
     }
     catch(err)
@@ -265,6 +309,319 @@ exports.addSubject=async(req,res)=>{
         
     }
 }
+
+
+exports.getAllSubjects=async(req,res)=>{
+    
+    try{
+        
+        const dId=req.teacher.department;
+        
+        const all=await subjectModel.find({department:dId});
+        
+        res.status(200).json({
+           status:"success",
+           data:all
+           
+             })
+        
+    }catch(err)
+    {
+        res.status(400).json({
+            status:"failure",
+            error:err.message
+        }) 
+    }
+    
+    
+}
+
+exports.deleteSubject=async(req,res)=>{
+    
+    try{
+        
+        const subId=req.params.id;
+        
+        const tId=req.teacher.department;
+        const subject=await subjectModel.findById(subId);
+        if(String(subject.department)==String(tId))
+        {
+            await subjectModel.deleteOne({_id:subId});    
+         res.status(201).json({
+           status:"success",
+           data:"subject deleted Successfully"
+           
+             })
+            
+            
+        }
+        else{
+            throw new Error("Your Not Authorized to delete it!!!");
+        }
+        
+    }
+    catch(err)
+    {
+        res.status(400).json({
+            status:"failure",
+            error:err.message
+        }) 
+    }
+}
+
+
+
+exports.addMeeting=async(req,res)=>{
+    
+   try{
+       
+       
+       const meeting=new meetingModel(req.body);
+       meeting.teacher=req.teacher._id;
+       meeting.section=req.params.secId;
+       meeting.subject=req.params.mId;
+      
+      
+      await meeting.save();
+      
+      res.status(201).json({
+           status:"success",
+           data:meeting
+           
+             })
+            
+       
+   } 
+   
+   catch(err)
+   {
+       res.status(400).json({
+            status:"failure",
+            error:err.message
+        }) 
+   }
+    
+    
+    
+    
+}
+
+exports.getAllMeetingsByTeacher=async(req,res)=>{
+    
+    try{
+        
+        const tId=req.teacher._id;
+        
+        
+        const all=await meetingModel.find({teacher:tId});
+        
+         res.status(201).json({
+           status:"success",
+           data:all
+           
+             })
+            
+        
+        
+        
+    }
+    catch(err)
+    {
+        res.status(400).json({
+            status:"failure",
+            error:err.message
+        }) 
+    }
+    
+}
+
+
+exports.deleteteacherLink=async(req,res)=>{
+    
+    try{
+        
+        const mId=req.params.id;
+        
+        const meeting=await meetingModel.findById(mId);
+        
+        if(String(meeting.teacher)==String(req.teacher._id))
+        {
+            await meetingModel.deleteOne({_id:mId});    
+            
+         res.status(201).json({
+           status:"success",
+           data:"meeting link deleted Successfully"
+           
+             })
+            
+        }
+        else{
+            throw new Error("Your not authorized to delete it");
+        }
+        
+        
+    }
+    
+    catch(err)
+    {
+     res.status(400).json({
+            status:"failure",
+            error:err.message
+        })    
+    }
+    
+    
+}
+
+
+
+
+exports.disactiveStudents=async(req,res)=>{
+    
+    try{
+        const dId=req.teacher.department;
+        const all=await studentModel.find({active:false,department:dId})
+         res.status(200).json({
+           status:"success",
+           data:all
+           
+             })
+        
+    }
+    
+    catch(err)
+    {
+        res.status(400).json({
+            status:"failure",
+            error:err.message
+        })   
+        
+    }
+    
+    
+    
+    
+}
+
+
+
+exports.getAllStudentsByDepartment=async(req,res)=>{
+    
+    
+    try{
+        
+        const dId=req.teacher.department;
+        
+        
+        const all=await studentModel.find({department:dId});
+         res.status(200).json({
+           status:"success",
+           data:all
+           
+             })
+        
+        
+        
+    }
+    catch(err)
+    {
+        res.status(400).json({
+            status:"failure",
+            error:err.message
+        })   
+        
+    }
+    
+}
+
+exports.makeStudentActive=async(req,res)=>{
+    
+    try{
+        
+        const sId=req.params.id;
+        
+        const student=await studentModel.findById(sId);
+        
+        if(student)
+        {
+            if(String(student.department)==String(req.teacher.department)){
+                
+                student.active=true
+                await student.save({validateBeforeSave:false});
+                
+                res.status(200).json({
+           status:"success",
+           data:student
+           
+             })
+            }
+            else{
+                throw new Error("Your Not Authorized To make active")
+            }
+            
+        }
+        else{
+            throw new Error("No Stundet Found")
+        }
+        
+        
+        
+    }
+    catch(err)
+    {
+     
+     res.status(400).json({
+            status:"failure",
+            error:err.message
+        })      
+    }
+    
+}
+
+
+
+
+exports.deleteStudent=async(req,res)=>{
+    
+    
+    try{
+        
+        const sId=req.params.id;
+        
+        const student=await studentModel.findById(sId);
+        
+        if(String(student.department)==String(req.teacher.department))
+        {
+            await studentModel.deleteOne({_id:sId});    
+            
+         res.status(201).json({
+           status:"success",
+           data:"student  deleted Successfully"
+           
+             })
+            
+        }
+        else{
+            throw new Error("Your not authorized to delete it");
+        }
+        
+        
+    }
+    catch(err)
+    {
+        
+        
+        res.status(400).json({
+            status:"failure",
+            error:err.message
+        })    
+    }
+    
+    
+    
+}
+
+
